+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- issues with empty or `null` labels in attribute export list
+- Shopgate configuration page > export attribute list update whenever SW `FreeText` attributes are manipulated
 
 ## [2.9.99] - 2021-09-08
 ### Fixed