+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Fixed
-- issues with empty or `null` labels in attribute export list
-- Shopgate configuration page > export attribute list update whenever SW `FreeText` attributes are manipulated
+- displaying empty or `null` labels in attribute export list
+- attribute export list in the plugin configuration not updating when `FreeText` attributes are changed
 
 ## [2.9.99] - 2021-09-08
 ### Fixed